+Added text rendering using an approach with bad performance (will fix in the next revision), but the interfaces won't change

+++ SimpleEngine/Src/SimpleSpriteSheet.cpp
@@ -38,7 +38,7 @@ void SimpleSpriteSheet::CreateUniformFrames(glm::ivec2 size, glm::ivec2 count) {
 
 glm::ivec4 SimpleSpriteSheet::GetCoordsForIndex(int idx) {
 
-	SIMPLE_ASSERT(idx < _frames.size());
+	SIMPLE_ASSERT(size_t(idx) < _frames.size());
 	
 	//Return the pixel coordinates for the current sheet
 	return _frames[idx];
